pyproject.toml
[tool.poetry]
name = "feedback_bot"
version = "0.0.2"
description = ""
authors = ["Nikita Churikov <nikita@chur.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.1"
loguru = "^0.3"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
run_feedback_bot = 'feedback_bot.main:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
